Fix bug for item shopping
If a user does not have enough money, the program
crashes. This happens because
"item_purchase_result.html" requires a parameter called
"already_has_item", and this parameter is
initialized inside an if statement that checks
if the user has enough money or if the user
is an admin. Move the function that sets
"already_has_item" out of the if statement. Break
this function into two parts because it both adds
an item to a user and also returns if it is already
owned by the user. Create a "check_item_ownership"
function to check if the user already owns the item.